# api/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: Service for storing metrics.
  title: MetricServer Endpoints
  version: "1.0"
paths:
...
  /update/counter/{name}/{value}:
    post:
      description: Update Counter metric
      operationId: updateCounter
      parameters:
      - description: Metric name
        in: path
        name: name
        required: true
        type: string
      - description: Metric value
        in: path
        name: value
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Wrong request, can't parse number
          schema:
            type: string
        "500":
          description: Internal error
          schema:
            type: string
      summary: Update Counters
      tags:
      - Update Metrics
  /update/gauge/{name}/{value}:
    post:
      description: Update gauge metric
      operationId: updateGauge
      parameters:
      - description: Metric name
        in: path
        name: name
        required: true
        type: string
      - description: Metric value
        in: path
        name: value
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Wrong request, can't parse number
          schema:
            type: string
        "500":
          description: Internal error
          schema:
            type: string
      summary: Update Gauges
      tags:
      - Update Metrics
...
swagger: "2.0"